Use 64bit types for memory addresses
mcelog reads 64-bit addresses from the MCE, but passes them around with
type unsigned long, which causes mcelog --dmi to give incorrect
results on 32-bit systems with more than 4 GiB of RAM.

See http://bugs.debian.org/437022.

+++ dimm.h
@@ -1,6 +1,6 @@
 void close_dimm_db(void);
 int open_dimm_db(char *fn);
-void new_error(unsigned long addr, unsigned long max_error, char *trigger);
+void new_error(unsigned long long addr, unsigned long max_error, char *trigger);
 void reset_dimm(char *locator);
 void gc_dimms(void);
 void dump_all_dimms(void);
